pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elevenlabs>=1.5.0",
    "fastapi==0.113.0",
    "google-api-core==2.17.1",
    "google-cloud-speech>=2.33.0",
    "google-cloud-texttospeech>=2.29.0",
    "google-generativeai==0.3.2",
    "gtts==2.3.2",
    "pydantic==2.6.2",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "soundfile==0.12.1",
    "uvicorn==0.23.2",
    "vosk==0.3.45",
]
